Reduce code duplication for pasting files in frontend
